+++ spec/lib/aixm/component/runway_spec.rb
@@ -52,51 +52,6 @@ describe AIXM::Component::Runway do
     end
   end
 
-  describe :composition= do
-    it "fails on invalid values" do
-      [:foobar, 123].wont_be_written_to subject, :composition
-    end
-
-    it "accepts nil value" do
-      [nil].must_be_written_to subject, :composition
-    end
-
-    it "looks up valid values" do
-      subject.tap { |s| s.composition = :macadam }.composition.must_equal :macadam
-      subject.tap { |s| s.composition = :GRADE }.composition.must_equal :graded_earth
-    end
-  end
-
-  describe :preparation= do
-    it "fails on invalid values" do
-      [:foobar, 123].wont_be_written_to subject, :preparation
-    end
-
-    it "accepts nil value" do
-      [nil].must_be_written_to subject, :preparation
-    end
-
-    it "looks up valid values" do
-      subject.tap { |s| s.preparation = :rolled }.preparation.must_equal :rolled
-      subject.tap { |s| s.preparation = :NATURAL }.preparation.must_equal :no_treatment
-    end
-  end
-
-  describe :condition= do
-    it "fails on invalid values" do
-      [:foobar, 123].wont_be_written_to subject, :condition
-    end
-
-    it "accepts nil value" do
-      [nil].must_be_written_to subject, :condition
-    end
-
-    it "looks up valid values" do
-      subject.tap { |s| s.condition = :fair }.condition.must_equal :fair
-      subject.tap { |s| s.condition = :GOOD }.condition.must_equal :good
-    end
-  end
-
   describe :status= do
     it "fails on invalid values" do
       [:foobar, 123].wont_be_written_to subject, :status
@@ -130,9 +85,15 @@ describe AIXM::Component::Runway do
           <valLen>650</valLen>
           <valWid>80</valWid>
           <uomDimRwy>M</uomDimRwy>
-          <codeComposition>GRADE</codeComposition>
-          <codePreparation>ROLLED</codePreparation>
-          <codeCondSfc>FAIR</codeCondSfc>
+          <codeComposition>ASPH</codeComposition>
+          <codePreparation>PAVED</codePreparation>
+          <codeCondSfc>GOOD</codeCondSfc>
+          <valPcnClass>59</valPcnClass>
+          <codePcnPavementType>F</codePcnPavementType>
+          <codePcnPavementSubgrade>A</codePcnPavementSubgrade>
+          <codePcnMaxTirePressure>W</codePcnMaxTirePressure>
+          <codePcnEvalMethod>T</codePcnEvalMethod>
+          <txtPcnNote>Paved shoulder on 2.5m on each side of the RWY.</txtPcnNote>
           <codeSts>CLSD</codeSts>
           <txtRmk>Markings eroded</txtRmk>
         </Rwy>
@@ -211,7 +172,8 @@ describe AIXM::Component::Runway do
 
     it "builds correct minimal OFMX" do
       AIXM.ofmx!
-      %i(length width composition preparation condition status remarks).each { |a| subject.send(:"#{a}=", nil) }
+      %i(length width status remarks).each { |a| subject.send(:"#{a}=", nil) }
+      %i(composition preparation condition pcn remarks).each { |a| subject.surface.send(:"#{a}=", nil) }
       subject.to_xml.must_equal <<~END
         <Rwy>
           <RwyUid>