+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 #### Additions
 * `Runway#preparation` and `Runway#condition`
+* Surface for `Runway|Helipad#surface`
 
 #### Breaking Changes
 * Renamed `AIXM::H` to `AIXM::A` (angle) and add simple arithmetics to make it more versatile
+* `Runway|Helipad#composition` moved to `Runway|Helipad#surface`
 
 ## 0.3.3